Ensure that the homepage does not load until the ticket subscription is returned
Otherwise the page will render with empty cards/no graphs because the tickets
array is empty when the home template renders.

Admittedly this fix is a bit of a hack (having hard-coded the default squad in
the route's 'waitOn' function). The reason for doing this is that the squad is
set in the client's session variable, which isn't available on the server.

We previously worked around this issue by calling the drawCharts() function in
the Tracker.autorun call, after the subscriptions were returned. Having
restructured the client-side code, we can no longer do that since the function
is defined in helpers.js.coffee, which loads after application.js.coffee.

Technically this hack will work, since the default value will always be
'Front End' and we only have to be concerned about this timing issue on initial
page load.