Ensure that cards reload when the tickets subscription is returned
The previous attempt to ensure that data is loaded only after the
ticket subscription is returned caused the subscription for the
Platform squad to return front end tickets.

This workaround essentially watches the ticket subscription and
reloads the charts when it is returned, which apparently reloads the
tabular cards as well.

Before trying this I explored the approach suggested in
http://stackoverflow.com/questions/20942025/waiting-for-meteor-collection-to-finish-before-next-step
but that caused a full template refresh when toggling squads, which
I felt detracted from the overall user experience.

This should not be considered a long-term fix for this issue.

One additional option, that I've been avoiding to this point, is to
track the selected squad in the route (rather than with a session variable)
and pass it to the subscription in IronRouters waitOn function. At
this point my preference is to avoid any route changes to keep the
current single-page app feel.