Integrate with the JIRA API.
The current authentication method relies on a user's JIRA username and password. In a
future version we'll look to implement a linked application that uses a unique token for
authentication.

The API calls rely on the presence of three filters - ids 11455, 10805, 11472. These
filters return all bugs, all non-bugs, and all tickets without components, respectively.
In a future version we'll look at passing these filters an argument for the selected
squad which would greatly reduce the number of tickets returned and eliminate the need to
filter by squad server-side.